// src/Modules/confirmEmail/__generated__/resendEmail.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: resendEmail
// ====================================================

export interface resendEmail {
  resendEmail: boolean | null;
}

export interface resendEmailVariables {
  email: string;
}